+++ CHANGELOG.md
@@ -34,6 +34,8 @@
   - Added `forced.pull`, `forced.push`, and `forced.slide` to adjust the distance of forced movement abilities. (#1321)
   - Added `keyword` to add keywords to abilities. (#1322)
   - Added `potency` to adjust the potency of abilities. (#1350)
+- Implemented a Reference page, which has additional room to specify a tooltip for the reference enricher. (#1423, #1425)
+- Added `system.artDescription` as a valid property for Image journal entry pages, providing rich alt text. (#1424)
 
 ### Changed